requirements.txt
langchain==0.3.26
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-chroma==0.1.4
langchain-google-genai<3.0.0
langchain-core<1.0.0
langchain-community==0.3.26
langchain-huggingface
flask